% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeValorisation.R
\name{datetimeValorisation}
\alias{datetimeValorisation}
\title{datetimeValorisation}
\usage{
datetimeValorisation(df, transformTimestamp = TRUE)
}
\arguments{
\item{df}{data frame containig the date/time field to be improve}

\item{transormTimestamp}{boolean value; TRUE if the timestamp field is in character format and need to be transformed}
}
\description{
Function to add key figures generated from a date/time field
}
\examples{
df <- datetimeValorisation(df, TRUE)
}

